split dispatch task script
